requirements.txt
adal==1.2.1
amqp==2.4.2
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.6.0.0
cachetools==3.1.1
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
click==6.7
cryptography==2.6.1
dataclasses==0.6
defusedxml==0.5.0
Django==2.0.3
django-braces==1.12.0
django-cors-headers==2.1.0
django-markdownx==2.0.28
django-modeltranslation==0.13.1
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
django-rest-hooks==1.5.0
djangorestframework==3.7.7
dnspython==1.16.0
greenlet==0.4.15
httplib2==0.12.3
idna==2.6
isodate==0.6.0
kombu==4.5.0
Markdown==3.1.1
mock==3.0.5
monotonic==1.5
msrest==0.6.6
msrestazure==0.6.0
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyJWT==1.5.3
pymongo==3.8.0
python-dateutil==2.8.0
python-dotenv==0.7.1
python3-openid==3.1.0
pytz==2017.3
redis==3.2.1
requests==2.18.4
requests-oauthlib==0.8.0
rsa==4.0
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
vine==1.3.0
